split __init__.py into mfk_parser.py and mfk_manager.py
